requirements.txt
Django~=4.2.1
psycopg2>=2.9.6
SQLAlchemy>=2.0.25
GeoAlchemy2==0.16.0
Fiona==1.9.4.post1
rtree==1.0.1
geopandas==0.13.0
djangorestframework~=3.14.0
djangorestframework-gis~=1.0.0
django-filter~=22.1.0
django-bootstrap5~=23.3
django-widget-tweaks~=1.4.8
celery
plotly
pylusat
